Adjust behavior of windowIsPopupType and hasTitleBar
The previous implementation did not check for the popup flag, so
it was added, and the tool window exception was preserved.
hasTitleBar was also changed so it checks for popups and not tooltips
specifically (tooltips are always popups).

Change-Id: I3e2ba3be56e992b30ca2a07375092073572e7fcb
Reviewed-by: Lorn Potter <lorn.potter@gmail.com>
(cherry picked from commit aa6405ee037abd53e88c7cd2ceee78fd973cafa9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>