Doc: Remove broken homepage link for Vera fonts attribution
The attribution for the Vera Fonts project declares a page on the gnome
domain as the homepage of the fonts.

The page is not maintained anymore and was removed from the gnome
domain.

Hence, we remove the "Homepage" key from the Vera Fonts attribution as
no homepage seems to exists at the moment.

Task-number: QTBUG-96127
Pick-to: 6.2
Change-Id: If4801eb96ce5568c4c18ecea7807ad87e165ca2a
Reviewed-by: Kai Koehne <kai.koehne@qt.io>